requirements.txt
# DataDeck - Automated PPT Report Generation System
# Core Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1

# PPT Generation
python-pptx==0.6.23
Pillow==10.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# AI/ML
openai==1.3.7
scikit-learn==1.3.2
nltk==3.8.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-slugify==8.0.1
pytz==2023.3.post1
email-validator==2.1.0

# Storage
boto3==1.29.7
minio==7.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

